+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/GovCert.hs
@@ -11,10 +11,10 @@
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
-module Cardano.Ledger.Conway.Rules.VDel (
-  ConwayVDEL,
-  ConwayVDelEvent (..),
-  ConwayVDelPredFailure (..),
+module Cardano.Ledger.Conway.Rules.GovCert (
+  ConwayGOVCERT,
+  ConwayGovCertEvent (..),
+  ConwayGovCertPredFailure (..),
 )
 where
 
@@ -25,7 +25,7 @@ import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..), encodeListLen)
 import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.CertState (VState (..))
 import Cardano.Ledger.Coin (Coin)
-import Cardano.Ledger.Conway.Era (ConwayVDEL)
+import Cardano.Ledger.Conway.Era (ConwayGOVCERT)
 import Cardano.Ledger.Conway.TxCert (ConwayCommitteeCert (..))
 import Cardano.Ledger.Core (Era (EraCrypto), EraPParams, EraRule, PParams)
 import Cardano.Ledger.Credential (Credential)
@@ -54,83 +54,83 @@ import Data.Word (Word8)
 import GHC.Generics (Generic)
 import NoThunks.Class (NoThunks (..))
 
-data ConwayVDelPredFailure era
-  = ConwayDRepAlreadyRegisteredVDEL !(Credential 'Voting (EraCrypto era))
-  | ConwayDRepNotRegisteredVDEL !(Credential 'Voting (EraCrypto era))
-  | ConwayDRepIncorrectDepositVDEL !Coin
-  | ConwayCommitteeHasResignedVDEL !(KeyHash 'CommitteeColdKey (EraCrypto era))
+data ConwayGovCertPredFailure era
+  = ConwayDRepAlreadyRegistered !(Credential 'Voting (EraCrypto era))
+  | ConwayDRepNotRegistered !(Credential 'Voting (EraCrypto era))
+  | ConwayDRepIncorrectDeposit !Coin
+  | ConwayCommitteeHasResigned !(KeyHash 'CommitteeColdKey (EraCrypto era))
   deriving (Show, Eq, Generic)
 
-instance NoThunks (ConwayVDelPredFailure era)
+instance NoThunks (ConwayGovCertPredFailure era)
 
-instance NFData (ConwayVDelPredFailure era)
+instance NFData (ConwayGovCertPredFailure era)
 
 instance
   (Typeable era, Crypto (EraCrypto era)) =>
-  EncCBOR (ConwayVDelPredFailure era)
+  EncCBOR (ConwayGovCertPredFailure era)
   where
   encCBOR = \case
-    ConwayDRepAlreadyRegisteredVDEL cred ->
+    ConwayDRepAlreadyRegistered cred ->
       encodeListLen 2
         <> encCBOR (0 :: Word8)
         <> encCBOR cred
-    ConwayDRepNotRegisteredVDEL cred ->
+    ConwayDRepNotRegistered cred ->
       encodeListLen 2
         <> encCBOR (1 :: Word8)
         <> encCBOR cred
-    ConwayDRepIncorrectDepositVDEL deposit ->
+    ConwayDRepIncorrectDeposit deposit ->
       encodeListLen 2
         <> encCBOR (2 :: Word8)
         <> encCBOR deposit
-    ConwayCommitteeHasResignedVDEL keyH ->
+    ConwayCommitteeHasResigned keyH ->
       encodeListLen 2
         <> encCBOR (3 :: Word8)
         <> encCBOR keyH
 
 instance
   (Typeable era, Crypto (EraCrypto era)) =>
-  DecCBOR (ConwayVDelPredFailure era)
+  DecCBOR (ConwayGovCertPredFailure era)
   where
-  decCBOR = decodeRecordSum "ConwayVDelPredFailure" $
+  decCBOR = decodeRecordSum "ConwayGovCertPredFailure" $
     \case
       0 -> do
         cred <- decCBOR
-        pure (2, ConwayDRepAlreadyRegisteredVDEL cred)
+        pure (2, ConwayDRepAlreadyRegistered cred)
       1 -> do
         cred <- decCBOR
-        pure (2, ConwayDRepNotRegisteredVDEL cred)
+        pure (2, ConwayDRepNotRegistered cred)
       2 -> do
         deposit <- decCBOR
-        pure (2, ConwayDRepIncorrectDepositVDEL deposit)
+        pure (2, ConwayDRepIncorrectDeposit deposit)
       3 -> do
         keyH <- decCBOR
-        pure (2, ConwayCommitteeHasResignedVDEL keyH)
+        pure (2, ConwayCommitteeHasResigned keyH)
       k -> invalidKey k
 
-newtype ConwayVDelEvent era = VDelEvent (Event (EraRule "VDEL" era))
+newtype ConwayGovCertEvent era = GovCertEvent (Event (EraRule "GOVCERT" era))
 
 instance
   ( EraPParams era
-  , State (EraRule "VDEL" era) ~ VState era
-  , Signal (EraRule "VDEL" era) ~ ConwayCommitteeCert (EraCrypto era)
-  , Environment (EraRule "VDEL" era) ~ PParams era
-  , EraRule "VDEL" era ~ ConwayVDEL era
-  , Eq (PredicateFailure (EraRule "VDEL" era))
-  , Show (PredicateFailure (EraRule "VDEL" era))
+  , State (EraRule "GOVCERT" era) ~ VState era
+  , Signal (EraRule "GOVCERT" era) ~ ConwayCommitteeCert (EraCrypto era)
+  , Environment (EraRule "GOVCERT" era) ~ PParams era
+  , EraRule "GOVCERT" era ~ ConwayGOVCERT era
+  , Eq (PredicateFailure (EraRule "GOVCERT" era))
+  , Show (PredicateFailure (EraRule "GOVCERT" era))
   ) =>
-  STS (ConwayVDEL era)
+  STS (ConwayGOVCERT era)
   where
-  type State (ConwayVDEL era) = VState era
-  type Signal (ConwayVDEL era) = ConwayCommitteeCert (EraCrypto era)
-  type Environment (ConwayVDEL era) = PParams era
-  type BaseM (ConwayVDEL era) = ShelleyBase
-  type PredicateFailure (ConwayVDEL era) = ConwayVDelPredFailure era
-  type Event (ConwayVDEL era) = ConwayVDelEvent era
+  type State (ConwayGOVCERT era) = VState era
+  type Signal (ConwayGOVCERT era) = ConwayCommitteeCert (EraCrypto era)
+  type Environment (ConwayGOVCERT era) = PParams era
+  type BaseM (ConwayGOVCERT era) = ShelleyBase
+  type PredicateFailure (ConwayGOVCERT era) = ConwayGovCertPredFailure era
+  type Event (ConwayGOVCERT era) = ConwayGovCertEvent era
 
-  transitionRules = [conwayVDelTransition @era]
+  transitionRules = [conwayGovCertTransition @era]
 
-conwayVDelTransition :: TransitionRule (ConwayVDEL era)
-conwayVDelTransition = do
+conwayGovCertTransition :: TransitionRule (ConwayGOVCERT era)
+conwayGovCertTransition = do
   TRC
     ( _pp
       , vState@VState {vsDReps, vsCommitteeHotKeys}
@@ -139,12 +139,12 @@ conwayVDelTransition = do
     judgmentContext
   case c of
     ConwayRegDRep cred _deposit -> do
-      Set.notMember cred vsDReps ?! ConwayDRepAlreadyRegisteredVDEL cred
+      Set.notMember cred vsDReps ?! ConwayDRepAlreadyRegistered cred
       -- TODO: check against a new PParam `drepDeposit`, once PParams are updated. -- someCheck ?! ConwayDRepIncorrectDeposit deposit
       pure $ vState {vsDReps = Set.insert cred vsDReps}
     ConwayUnRegDRep cred _deposit -> do
       -- TODO: check against a new PParam `drepDeposit`, once PParams are updated. -- someCheck ?! ConwayDRepIncorrectDeposit deposit
-      Set.member cred vsDReps ?! ConwayDRepNotRegisteredVDEL cred
+      Set.member cred vsDReps ?! ConwayDRepNotRegistered cred
       pure $ vState {vsDReps = Set.delete cred vsDReps}
     ConwayAuthCommitteeHotKey coldK hotK -> do
       checkColdKeyHasNotResigned coldK vsCommitteeHotKeys
@@ -156,4 +156,4 @@ conwayVDelTransition = do
     checkColdKeyHasNotResigned coldK vsCommitteeHotKeys =
       (isNothing <$> Map.lookup coldK vsCommitteeHotKeys)
         /= Just True
-        ?! ConwayCommitteeHasResignedVDEL coldK
+        ?! ConwayCommitteeHasResigned coldK